fix: move dhlab_v1 code into its own subpackage
